requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
openai==1.3.0
Pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
Werkzeug==2.3.6